add motor direction, rename power to motor_power
